@@ -9,6 +9,7 @@ import {ArticleDetailComponent} from '../article-detail/article-detail.component
 import {UserCenterComponent} from '../user-center/user-center.component';
 import {LogoutComponent} from '../logout/logout.component';
 import {ArticlePublishComponent} from '../article-publish/article-publish.component';
+import {AskComponent} from '../ask/ask.component';
 
 const routes: Routes = [
   // 默认路由，这个路由会把一个与空路径"完全匹配"的 URL 重定向到路径为 '/index' 的路由
@@ -20,7 +21,8 @@ const routes: Routes = [
   {path: 'article/detail/:id', component: ArticleDetailComponent},
   {path: 'article/write', component: ArticlePublishComponent},
   {path: 'user/center/:id', component: UserCenterComponent},
-  {path: 'user/logout', component: LogoutComponent}
+  {path: 'user/logout', component: LogoutComponent},
+  {path: 'ask', component: AskComponent}
 ];
 
 @NgModule({
